software-design/05-drawing/requirements.txt
pygame
pycairo
matplotlib
numpy
